Spectrogram code created
Notebook extracts 224 and 512 spects. 15s chunks with 25% overlap.

Tests to determine fft dim. and hop size were done and found the existing values to be the best for the 16kHz mbari files ( in terms of quality and sharpness of the spects, and efficiency. smaller hops were tested, no perceived diff in quality, but slows the process considerably)

The code normalizes each chunk independently to minimize the effect of random loud bumps into the hydrophone.